successfuly delete documents
